feat: Room with manual DI
