+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
 
   <application
       android:allowBackup="true"
+      android:name=".TodoApplication"
       android:dataExtractionRules="@xml/data_extraction_rules"
       android:fullBackupContent="@xml/backup_rules"
       android:icon="@mipmap/ic_launcher"
