+++ app/src/main/java/com/juandgaines/todoapp/TodoApplication.kt
@@ -1,0 +1,12 @@
+package com.juandgaines.todoapp
+
+import android.app.Application
+import com.juandgaines.todoapp.data.DataSourceFactory
+import com.juandgaines.todoapp.domain.TaskLocalDataSource
+
+class TodoApplication:Application() {
+
+    val dataSource :TaskLocalDataSource
+        get() = DataSourceFactory.createDataSource(this)
+
+}
