+++ app/src/main/java/com/juandgaines/todoapp/data/RoomTaskLocalDataSource.kt
@@ -2,38 +2,43 @@ package com.juandgaines.todoapp.data
 
 import com.juandgaines.todoapp.domain.Task
 import com.juandgaines.todoapp.domain.TaskLocalDataSource
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.withContext
 
 class RoomTaskLocalDataSource(
-    private val taskDao: TaskDao
+    private val taskDao: TaskDao,
+    private val dispatcherIO: CoroutineDispatcher = Dispatchers.IO
 ):TaskLocalDataSource {
     override val tasksFlow: Flow<List<Task>>
         get() = taskDao.getAllTasks().map {
             it.map { taskEntity -> taskEntity.toTask() }
-        }
+        }.flowOn(dispatcherIO)
 
-    override suspend fun addTask(task: Task) {
+    override suspend fun addTask(task: Task) = withContext(dispatcherIO) {
         taskDao.upsertTask(TaskEntity.fromTask(task))
     }
 
-    override suspend fun updateTask(updatedTask: Task) {
+    override suspend fun updateTask(updatedTask: Task) = withContext(dispatcherIO) {
         taskDao.upsertTask(TaskEntity.fromTask(updatedTask))
     }
 
-    override suspend fun removeTask(task: Task) {
+    override suspend fun removeTask(task: Task) = withContext(dispatcherIO) {
         taskDao.deleteTaskById(task.id)
     }
 
-    override suspend fun deleteAllTasks() {
+    override suspend fun deleteAllTasks() = withContext(dispatcherIO) {
         taskDao.deleteAllTasks()
     }
 
-    override suspend fun getTaskById(taskId: String): Task? {
-        return taskDao.getTaskById(taskId)?.toTask()
+    override suspend fun getTaskById(taskId: String): Task? = withContext(dispatcherIO) {
+       taskDao.getTaskById(taskId)?.toTask()
     }
 
-    override suspend fun removeAllTasks() {
+    override suspend fun removeAllTasks() = withContext(dispatcherIO){
         taskDao.deleteAllTasks()
     }
 }
